Update MK1.py
Moved the constants to the top of the script. Easier to follow your code. Further, add a general definition of each class and the role it plays.
Especially if you are going to have other people join the project you'll need to organize and comment a lot more. Also, try to keep your variable names more specific because with so many different parts, its easier to follow, for both you and us (the readers)

Overall, it will be much easier to change and add on if you make it easier to read. Hope this helps.
I can do some of it, but i dont know what is what so its hard to change up.